backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
firebase-admin==6.5.0
groq==0.11.0
google-generativeai==0.8.0
anthropic==0.34.0
feedparser==6.0.11
python-dotenv==1.0.1
